corregi los errores del desafio complementario y agregue validaciones para los metodos 'put' y 'delete' del carrito

+++ pre-entrega-2/src/routes/cart.router.js
@@ -19,7 +19,7 @@ module.exports = (cartManager, productManager) => {
         try {
             const cart = await cartManager.getCartById(cartId)
             if (!cart) {
-                res.status(400).json({ error: `No cart exists with the id ${cartId}` })
+                res.status(404).json({ error: `No cart exists with the id ${cartId}` })
             } else {
                 res.json(cart.products)
             }
@@ -38,15 +38,21 @@ module.exports = (cartManager, productManager) => {
             // Se verifica si el carrito existe en el array de carritos
             const verifyCartId = await cartManager.getCartById(cartId)
             if (!verifyCartId) {
-                res.status(400).json({ error: `No cart exists with the id ${cartId}` })
+                res.status(404).json({ error: `No cart exists with the id ${cartId}` })
                 return cartId
             }
 
             // Se verifica si el producto existe en el array de productos
             const verifyProductId = await productManager.getProductById(productId)
             if (!verifyProductId) {
-                res.status(400).json({ error: `A product with the id ${productId} was not found.` })
+                res.status(404).json({ error: `A product with the id ${productId} was not found.` })
                 return productId
+            }
+
+            // Se verifica que la cantidad sea un número positivo
+            if (typeof quantity !== 'number' || quantity <= 0) {
+                res.status(404).json({ error: `Quantity (${quantity}) must be a positive number.` })
+                return quantity
             }
 
             const updateCart = await cartManager.addProductToCart(cartId, productId, quantity, productManager)
@@ -57,15 +63,13 @@ module.exports = (cartManager, productManager) => {
         }
     })
 
-
-
     // Ruta DELETE /api/carts/:cid - se eliminan todos los productos del carrito (se vacía el mismo)
     router.delete('/:cid', async (req, res) => {
         const cartId = req.params.cid
         try {
-            // se llama a la función clearCart del cartManager para eliminar todos los productos del carrito
+            // Se llama a la función clearCart del cartManager para eliminar todos los productos del carrito
             await cartManager.clearCart(cartId)
-            // se envía una respuesta exitosa al cliente
+            // Se envía una respuesta exitosa al cliente
             res.status(200).json({ message: 'All products deleted from cart successfully.' })
         } catch (error) {
             console.error("Error deleting products from cart", error)
@@ -77,15 +81,21 @@ module.exports = (cartManager, productManager) => {
     router.delete('/:cid/product/:pid', async (req, res) => {
         const cartId = req.params.cid
         const productId = req.params.pid
-
         try {
-            // Verificar si el carrito existe
+            // Se verifica si el carrito existe en el array de carritos
             const cart = await cartManager.getCartById(cartId)
             if (!cart) {
                 return res.status(404).json({ error: `Cart with id ${cartId} not found` })
             }
 
-            // Eliminar el producto del carrito
+            // Se verifica si el producto existe en el array de productos
+            const verifyProductId = await productManager.getProductById(productId)
+            if (!verifyProductId) {
+                res.status(404).json({ error: `A product with the id ${productId} was not found.` })
+                return productId
+            }
+
+            // Se elimina el producto del carrito
             await cartManager.deleteProductFromCart(cartId, productId)
             res.json({ message: `Product with id ${productId} removed from cart with id ${cartId}` })
         } catch (error) {
@@ -96,15 +106,40 @@ module.exports = (cartManager, productManager) => {
 
     // Ruta PUT api/carts/:cid/products/:pid - se actualiza la cantidad de un producto especifico en el carrito
     router.put('/:cid/product/:pid', async (req, res) => {
+        const cartId = req.params.cid
+        const productId = req.params.pid
+        const { quantity } = req.body
         try {
-            const cartId = req.params.cid
-            const productId = req.params.pid
-            const { quantity } = req.body
+            // Se verifica si el carrito existe en el array de carritos
+            const cart = await cartManager.getCartById(cartId)
+            if (!cart) {
+                return res.status(404).json({ error: `Cart with id ${cartId} not found` })
+            }
+
+            // Se verifica si el producto existe en el array de productos
+            const verifyProductId = await productManager.getProductById(productId)
+            if (!verifyProductId) {
+                res.status(404).json({ error: `A product with the id ${productId} was not found.` })
+                return productId
+            }
+
+            // se verifica si el producto existe en el carrito
+            const productToUpdate = cart.products.find(p => p.product.equals(productId))
+            if (!productToUpdate) {
+                res.status(404).json({ error: `A product with the id ${productId} was not found in the cart.` })
+                return null
+            }
+
+            // Se verifica que la cantidad sea un número positivo
+            if (typeof quantity !== 'number' || quantity <= 0) {
+                res.status(404).json({ error: `Quantity (${quantity}) must be a positive number.` })
+                return quantity
+            }
 
             // Se actualiza la cantidad del producto en el carrito
             const updatedCart = await cartManager.updateProductQuantityInCart(cartId, productId, quantity)
 
-            // Se Envia una respuesta con el carrito actualizado
+            // Se envia una respuesta con el carrito actualizado
             res.json(updatedCart)
         } catch (error) {
             console.error("Error updating product quantity in cart", error)
@@ -114,17 +149,16 @@ module.exports = (cartManager, productManager) => {
 
     // Ruta PUT api/carts/:cid - se actualiza el carrito con un arreglo de productos
     router.put('/:cid', async (req, res) => {
-        const cartId = req.params.cid;
-        const newProducts = req.body.products;
-
+        const cartId = req.params.cid
+        const newProducts = req.body.products
         try {
-            // se verifica si el carrito existe en el array de carritos
+            // Se verifica si el carrito existe en el array de carritos
             const cart = await cartManager.getCartById(cartId)
             if (!cart) {
                 return res.status(404).json({ error: `Cart with id ${cartId} not found` })
             }
 
-            // se actualiza el carrito con los nuevos productos
+            // Se actualiza el carrito con los nuevos productos
             const updatedCart = await cartManager.updateCart(cartId, newProducts)
 
             res.json(updatedCart)
